Abstracting VS Code specific code and generally tidying up shop.
