+++ package.json
@@ -16,7 +16,7 @@
 	},
 	"license": "MIT",
 	"publisher": "goblin",
-	"version": "0.2.3",
+	"version": "0.2.4",
 	"engines": {
 		"vscode": "^1.61.0"
 	},
@@ -33,7 +33,7 @@
 		"onCommand:rmg.create",
 		"onCommand:rmg.add"
 	],
-	"main": "./dist/extension.js",
+	"main": "./dist/vscode.js",
 	"contributes": {
 		"configuration": {
 			"title": "RMG",
@@ -347,12 +347,12 @@
 	"scripts": {
 		"vscode:prepublish": "npm run package",
 		"debug": "node ./node_modules/webpack/bin/webpack --watch",
-		"compile": "webpack",
+		"build": "webpack",
 		"watch": "webpack --watch",
 		"package": "webpack --mode production --devtool hidden-source-map",
-		"compile-tests": "tsc -p . --outDir out",
+		"build-tests": "tsc -p . --outDir out",
 		"watch-tests": "tsc -p . -w --outDir out",
-		"pretest": "npm run compile-tests && npm run compile && npm run lint",
+		"pretest": "npm run build-tests && npm run build && npm run lint",
 		"lint": "eslint src --ext ts",
 		"test": "node ./out/test/runTest.js"
 	},
