+++ src/templates/index.ts
@@ -1,0 +1,7 @@
+export * from "./template/base";
+export * from "./template/barrel";
+export * from "./template/component";
+export * from "./template/style";
+export * from "./template/test";
+export * from "./template/translation";
+export * from "./validator";
